ACTIN-643: Move medication interpretation status files to common
