+++ algo/src/main/kotlin/com/hartwig/actin/algo/evaluation/washout/HasRecentlyReceivedCancerTherapyOfName.kt
@@ -5,8 +5,8 @@ import com.hartwig.actin.algo.datamodel.Evaluation
 import com.hartwig.actin.algo.evaluation.EvaluationFactory
 import com.hartwig.actin.algo.evaluation.EvaluationFunction
 import com.hartwig.actin.algo.evaluation.util.Format.concat
-import com.hartwig.actin.algo.medication.MedicationStatusInterpretation
-import com.hartwig.actin.algo.medication.MedicationStatusInterpreter
+import com.hartwig.actin.clinical.interpretation.MedicationStatusInterpretation
+import com.hartwig.actin.clinical.interpretation.MedicationStatusInterpreter
 import com.hartwig.actin.clinical.datamodel.Medication
 
 class HasRecentlyReceivedCancerTherapyOfName(
