+++ algo/src/test/kotlin/com/hartwig/actin/algo/evaluation/medication/MedicationTestFactory.kt
@@ -2,8 +2,8 @@ package com.hartwig.actin.algo.evaluation.medication
 
 import com.hartwig.actin.PatientRecord
 import com.hartwig.actin.TestDataFactory
-import com.hartwig.actin.algo.medication.MedicationStatusInterpretation
-import com.hartwig.actin.algo.medication.MedicationStatusInterpreter
+import com.hartwig.actin.clinical.interpretation.MedicationStatusInterpretation
+import com.hartwig.actin.clinical.interpretation.MedicationStatusInterpreter
 import com.hartwig.actin.clinical.datamodel.AtcClassification
 import com.hartwig.actin.clinical.datamodel.CypInteraction
 import com.hartwig.actin.clinical.datamodel.Dosage
