+++ common/src/main/kotlin/com/hartwig/actin/clinical/interpretation/MedicationStatusInterpretation.kt
@@ -1,4 +1,4 @@
-package com.hartwig.actin.algo.medication
+package com.hartwig.actin.clinical.interpretation
 
 enum class MedicationStatusInterpretation {
     ACTIVE, PLANNED, STOPPED, CANCELLED, UNKNOWN
