--- report/src/main/kotlin/com/hartwig/actin/report/pdf/tables/clinical/MedicationStatusInterpretation.kt
@@ -1,6 +1,0 @@
-package com.hartwig.actin.report.pdf.tables.clinical
-
-// TODO (CB): move this class and the copy in algo and the test files to common module once common is converted to kotlin
-enum class MedicationStatusInterpretation {
-    ACTIVE, PLANNED, STOPPED, CANCELLED, UNKNOWN
-}
